Made the change to use python instead of shell
So my initial thought was to provide the utils in shell format, as that
would require 0 dependincies on most dev machines (except windows which
would require WSL). Instead I opted to develop the utils in python.

This is for multiple reasons. Mostly for my own sanity, but also since
the target audiance is python and jupyter users. Which means they would
already have python insalled. I will be using only system provided
packages and keep the code as backwards compatible as possible.

As such, the switch to python as the tool language has been made.