Fixed bug in intermediate release workflow
So I noticed that the intermediate release workflow only released patch
version, while I was sure that according to the logic it should have
brought out at least one minor version update.

I created a local test shell script and ran it for a commit of which I
new should have been a minor version bump (02e8151). The changes made to
the workflow fixed the logic.