Refactor file and folder names in .dsutils
The commit refactors the names of files and folders in the .dsutils directory to improve clarity and consistency.

Since the last commit I have opted to use the environment file for storing constents. That is, where the project resonates and the relative files and folders to it.

I also opted to steer away from usig a artifacts.py file and instead incorperate a looped import mechanism in experiments notebooks which import all files from the artifacts folder. As such, the artifacts.py is ommited from the files_and_folders.json